requirements.txt
streamlit==1.32.2
polygon-api-client==1.13.4
bokeh==2.4.3